Fix home page not marked as active in menu
Previously, with a menu item as follows, the "Home" link did not get
the `active` CSS class and was thus not styled as such.

```toml
[[menu.main]]
name = "Home"
url = "/"
weight = 1
```

The approach used here to determine whether the current page
corresponds to a certain menu entry was inspired by the one suggested
at https://stackoverflow.com/a/56454338. It is supposedly also more
robust when dealing with multilingual sites.